--- a/test/interaction/Issue6787-2.out
+++ b/test/interaction/Issue6787-2.out
@@ -6,5 +6,5 @@
 ((last . 1) . (agda2-goals-action '(0)))
 (agda2-give-action 0 "refl")
 (agda2-status-action "")
-(agda2-info-action "*All Warnings*" "Issue6787-2.agda:6.7-11: warning: -W[no]UnsupportedIndexedMatch This clause uses pattern-matching features that are not yet supported by Cubical Agda, the function to which it belongs will not compute when applied to transports. Reason: it relies on higher-dimensional unification, which is not yet supported when checking that the expression ? has type refl ≡ refl" nil)
+(agda2-info-action "*All Warnings*" "Issue6787-2.agda:6.7-11: warning: -W[no]UnsupportedIndexedMatch This clause uses pattern-matching features that are not yet supported by Cubical Agda, the function to which it belongs will not compute when applied to transports. Reason: It relies on injectivity of the data constructor refl, which is not yet supported when checking that the expression ? has type refl ≡ refl" nil)
 ((last . 1) . (agda2-goals-action '()))