- Remove unnecessary left inverse computations.
  - Previously, when checking clause bodies, the left inverses were computed and immediately discarded. Now they are not computed.
  - Previously, if a `UnifyLog` contained an unsupported step in left inversion, we would fully compute the left inverse from the log until we hit the offending step. Now, we first detect unsupported steps in the log and only build actual terms after that. 
- Rearrange benchmarking for `UnifyIndices`. `UnifyIndices` can be called from `Typing` and from `Coverage` and used to belong to both. Now it is its own entry and also has left inverse building as a sub-entry. I don't have strong feelings about the `UnifyIndices` change, but the addition of the left inversion benchmark is IMO well justified. 

`std-lib` times:

|      | master | PR |
|-----|------------|------|
| Total | 140,088ms | 135,489ms |
| UnifyIndices (total) | 14,010ms | 8,804ms |
| Total alloc | 607.4G | 573.4G |

Note that the left inverses which are actually used, are still computed, so most of the overhead of left inversion is still present.