@AndrasKovacs 2025-09-16
Some failing tests: https://github.com/AndrasKovacs/agda/pull/3