@AndrasKovacs 2025-09-18
CI passed: https://github.com/AndrasKovacs/agda/pull/3