@AndrasKovacs 2025-09-24
@jespercockx Here it's not a problem, at the call sites we either immediately run the `TCM` or it immediately becomes garbage. There's no strong reason to use `TCM`, I only used it because it seemed like the minimal code change.