@andreasabel 2025-11-01
@AndrasKovacs Not sure whether it is related to this PR, but `test/Fail/Issue1075` got 30% slower from 2.8.0 to current master:
```
$ agda-2.8.0 Issue1075.agda --profile=internal

Total                        634ms
Miscellaneous                 19ms
Coverage                      28ms (222ms)
Coverage.UnifyIndices        193ms
Typing                        30ms (214ms)
Typing.CheckLHS               38ms (107ms)
Typing.CheckLHS.UnifyIndices  69ms
Typing.CheckRHS               49ms
Typing.TypeSig                20ms
Termination                    9ms  (89ms)
Termination.Graph             73ms
Parsing                       19ms  (49ms)
Parsing.OperatorsExpr         24ms
Deserialization               35ms  (40ms)
Highlighting                  11ms
Positivity                    11ms

0.66s user
0.04s system 98% cpu
0.707 total

$ agda Issue1075.agda --profile=internal  

Total             811ms
Miscellaneous      21ms
UnifyIndices      424ms
Typing             43ms (246ms)
Typing.CheckRHS    83ms
Typing.CheckLHS    70ms
Typing.TypeSig     35ms
Coverage           38ms
Termination         7ms  (34ms)
Termination.Graph  20ms
Deserialization    32ms
Parsing            17ms  (24ms)
Positivity         15ms

0.85s user 
0.04s system 98% cpu 
0.899 total
```
More than 50% (424/811) spent on `UnifyIndices`.  Before (2.8.0) it was 40% (69 + 193)/634.