@AndrasKovacs 2025-11-01
@andreasabel It's unlikely to be related. The test doesn't use cubical or cubical-compatible, so this PR shouldn't affect it.